requirements.txt
asgiref==3.4.1
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
backports.zoneinfo==0.2.1
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.1
cycler==0.11.0
daphne==3.0.2
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
fonttools==4.32.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.3
incremental==21.3.0
kiwisolver==1.4.2
Markdown==3.3.6
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==21.7.0
txaio==21.2.1
typing_extensions==4.0.1
tzdata==2021.5
zipp==3.8.0
zope.interface==5.4.0